Add basic system allowing bots to switch weapons
Which weapon is more useful is checked by comparing the possible damage output and if not on cooldown, weapons are switched accordingly.
So far this is only a framework, since bots can only be given one weapon by default and their secondary is an NPC weapon which will almost always have better stats.