Improve bot revive objective
For manual revives the bot with the shortest path to the downed player is chosen instead of the bot with the shortest direct distance.
If Inspire is available and a bot is in range to inspire, that bot is chosen to revive, independent of path length.
Bot Inspire usage also does not depend on their detected objects anymore so it should be a lot more consistent.